update null test confirmed

+++ Source/UI/PickerPanel.h
@@ -162,7 +162,7 @@ public:
         }
         else if (item.type == 0 && item.audioRef != nullptr) {
             double miniZoom = (double)clipRect.getWidth() / std::max(0.1, (double)item.audioRef->width);
-            WaveformRenderer::drawWaveform(g, *item.audioRef, clipRect, item.color, WaveformViewMode::Combined, miniZoom, 0.0, 0.0);
+            WaveformRenderer::drawWaveform(g, *item.audioRef, clipRect.toFloat(), item.color, WaveformViewMode::Combined, miniZoom, 0.0, 0.0);
         }
         else if (item.type == 2 && item.autoRef != nullptr) {
             g.setColour(item.color.brighter());
